requirements.txt
flask==3.0.2
pandas==2.2.1
boto3==1.34.34
python-dotenv==1.0.1
requests==2.31.0 